feat(rdr3): 1355 support for Epic version (tested with RDO)

+++ code/components/ros-patches-five/src/AccountID.cpp
@@ -36,17 +36,26 @@ struct ExternalROSBlob
 	uint8_t data[16384];
 	uint8_t steamData[64 * 1024];
 	size_t steamSize;
+	uint8_t epicData[64 * 1024];
+	size_t epicSize;
 	uint32_t steamAppId;
 	bool valid;
 	bool tried;
+	bool triedEpic;
 
 	ExternalROSBlob()
 	{
 		valid = false;
 		tried = false;
+		triedEpic = false;
 		steamAppId = 0;
 		memset(data, 0, sizeof(data));
+
 		memset(steamData, 0, sizeof(steamData));
+		steamSize = 0;
+
+		memset(epicData, 0, sizeof(epicData));
+		epicSize = 0;
 	}
 };
 
@@ -180,6 +189,352 @@ static bool InitAccountSteam()
 	return blob->valid;
 }
 
+static bool InitAccountEOS()
+{
+	if (GetFileAttributesW(MakeRelativeGamePath(L"EOSSDK-Win64-Shipping.dll").c_str()) == INVALID_FILE_ATTRIBUTES)
+	{
+		return false;
+	}
+
+	static HostSharedData<ExternalROSBlob> blob("Cfx_ExtRosBlob");
+
+	if (!blob->triedEpic)
+	{
+#ifndef GTA_FIVE
+		if (!getenv("CitizenFX_ToolMode"))
+#endif
+		{
+			RunLauncher(L"ros:epic", true);
+		}
+	}
+
+	if (blob->valid)
+	{
+		accountBlob = blob->data;
+		trace("Epic ROS says it's signed in: %s\n", (const char*)&accountBlob[8]);
+	}
+
+	return blob->valid;
+}
+
+HANDLE OpenProcessByName(const std::wstring& name, DWORD desiredAccess)
+{
+	std::vector<DWORD> processes(16384);
+	DWORD cbNeeded = 0;
+	EnumProcesses(&processes[0], processes.size() * sizeof(DWORD), &cbNeeded);
+
+	auto numProcs = cbNeeded / sizeof(DWORD);
+
+	for (size_t i = 0; i < numProcs; i++)
+	{
+		HANDLE handle = OpenProcess(PROCESS_QUERY_LIMITED_INFORMATION, FALSE, processes[i]);
+
+		if (handle)
+		{
+			wchar_t filename[2048];
+			if (GetProcessImageFileNameW(handle, filename, std::size(filename)))
+			{
+				if (wcsstr(filename, name.c_str()))
+				{
+					CloseHandle(handle);
+
+					return OpenProcess(desiredAccess, FALSE, processes[i]);
+				}
+			}
+
+			CloseHandle(handle);
+		}
+	}
+
+	return INVALID_HANDLE_VALUE;
+}
+
+#include <shellapi.h>
+#include <winternl.h>
+
+#pragma comment(lib, "ntdll.lib")
+
+#include <eos-sdk/eos_sdk.h>
+#include <eos-sdk/eos_auth.h>
+
+void ValidateEpic(int parentPid)
+{
+	static HostSharedData<ExternalROSBlob> blob("Cfx_ExtRosBlob");
+	blob->triedEpic = true;
+
+	// find a process named EpicGamesLauncher.exe
+	auto egl = OpenProcessByName(L"EpicGamesLauncher.exe", PROCESS_VM_READ);
+
+	if (egl == INVALID_HANDLE_VALUE)
+	{
+		return;
+	}
+
+	if (!egl)
+	{
+		MessageBoxW(NULL, L"Could not read data from the Epic Games Launcher. If you are running it as administrator, please launch it as regular user.", L"RedM", MB_OK | MB_ICONSTOP);
+		return;
+	}
+
+	CloseHandle(egl);
+
+	// try finding MTL
+	auto mtl = OpenProcessByName(L"Launcher\\Launcher.exe", SYNCHRONIZE);
+
+	if (mtl != INVALID_HANDLE_VALUE)
+	{
+		MessageBoxW(NULL, L"If you're trying to launch an Epic Games version of RDR2, please make sure the Rockstar Games Launcher is closed. If not, remove EOSSDK-Win64-Shipping.dll from your RDR2 folder.", L"RedM", MB_OK | MB_ICONSTOP);
+		return;
+	}
+
+	// try to launch RDR-MTL from EGL
+	ShellExecuteW(NULL, L"open", L"com.epicgames.launcher://apps/Heather?action=launch&silent=true", NULL, NULL, SW_SHOWNORMAL);
+
+	// wait for MTL to open
+	int iterations = 0;
+
+	mtl = INVALID_HANDLE_VALUE;
+
+	while (mtl == INVALID_HANDLE_VALUE)
+	{
+		if (iterations > (60000 / 500))
+		{
+			MessageBoxW(NULL, L"If you're trying to launch an Epic Games version of RDR2, please make sure that you can launch it using the Epic Games Launcher first.", L"RedM", MB_OK | MB_ICONSTOP);
+			return;
+		}
+
+		Sleep(500);
+
+		mtl = OpenProcessByName(L"Launcher\\Launcher.exe", PROCESS_VM_READ | PROCESS_QUERY_INFORMATION | PROCESS_TERMINATE | PROCESS_SUSPEND_RESUME);
+	}
+
+	if (!mtl)
+	{
+		FatalError("Epic Games Launcher launched the Rockstar Games Launcher in a way we couldn't access. Aborting!");
+	}
+
+	// halt!
+	DebugActiveProcess(GetProcessId(mtl));
+
+	auto bye = [mtl]()
+	{
+		DebugActiveProcessStop(GetProcessId(mtl));
+		TerminateProcess(mtl, 0);
+	};
+
+	// get command line
+	PROCESS_BASIC_INFORMATION pbi;
+	NtQueryInformationProcess(mtl, ProcessBasicInformation, &pbi, sizeof(pbi), NULL);
+
+	PVOID pPEB = pbi.PebBaseAddress;
+	PVOID rtlUserProcParamsAddress;
+	UNICODE_STRING commandLine;
+
+	ReadProcessMemory(mtl, (PCHAR)pPEB + offsetof(PEB, ProcessParameters), &rtlUserProcParamsAddress, sizeof(PVOID), NULL);
+	ReadProcessMemory(mtl, (PCHAR)rtlUserProcParamsAddress + offsetof(RTL_USER_PROCESS_PARAMETERS, CommandLine), &commandLine, sizeof(commandLine), NULL);
+
+	std::vector<wchar_t> buffer((commandLine.Length / 2) + 1);
+	ReadProcessMemory(mtl, commandLine.Buffer, &buffer[0], commandLine.Length, NULL);
+
+	std::wstring commandLineString = &buffer[0];
+
+	int argc = 0;
+	auto argv = CommandLineToArgvW(commandLineString.c_str(), &argc);
+
+	std::wstring epicPassword;
+
+	for (size_t i = 0; i < argc; i++)
+	{
+		std::wstring_view arg = argv[i];
+
+		if (arg.find(L"-AUTH_PASSWORD=") == 0)
+		{
+			epicPassword = arg.substr(15);
+		}
+	}
+
+	LocalFree(argv);
+
+	if (epicPassword.empty())
+	{
+		FatalError("Didn't get an Epic exchange token.");
+	}
+
+	auto eosDll = LoadLibraryW(MakeRelativeGamePath(L"EOSSDK-Win64-Shipping.dll").c_str());
+
+	if (!eosDll)
+	{
+		FatalError("No Epic DLL.");
+	}
+
+	EOS_InitializeOptions eio = { 0 };
+	eio.ApiVersion = 2; //EOS_INITIALIZE_API_LATEST;
+	eio.ProductName = "RDR2";
+	eio.ProductVersion = "1.0";
+
+	auto _EOS_Initialize = (decltype(&EOS_Initialize))GetProcAddress(eosDll, "EOS_Initialize");
+	_EOS_Initialize(&eio);
+
+	EOS_Platform_Options epo = { 0 };
+	epo.ApiVersion = 5; //EOS_PLATFORM_OPTIONS_API_LATEST;
+	epo.ProductId = "1893447b7ea5459b96191a7458075768";
+	epo.SandboxId = "b30b6d1b4dfd4dcc93b5490be5e094e5";
+	epo.ClientCredentials.ClientId = "6b3b277c50a74855ab9beff2d62428aa";
+	epo.ClientCredentials.ClientSecret = "4h8MBChGbrV2woGgr2fdaYn2iWtHdf9F"; // omg this is SO secret, wow
+	epo.DeploymentId = "a59e364fb7f04ea4b0ca52e91281b9dc";
+	epo.Flags = EOS_PF_DISABLE_OVERLAY;
+
+	auto _EOS_Platform_Create = (decltype(&EOS_Platform_Create))GetProcAddress(eosDll, "EOS_Platform_Create");
+	auto epicPlatform = _EOS_Platform_Create(&epo);
+	assert(epicPlatform);
+
+	auto _EOS_Platform_GetAuthInterface = (decltype(&EOS_Platform_GetAuthInterface))GetProcAddress(eosDll, "EOS_Platform_GetAuthInterface");
+	static auto epicAuth = _EOS_Platform_GetAuthInterface(epicPlatform);
+
+	auto nepicPassword = ToNarrow(epicPassword);
+
+	EOS_Auth_Credentials ecs = { 0 };
+	ecs.ApiVersion = 1; //EOS_AUTH_CREDENTIALS_API_LATEST;
+	ecs.Type = EOS_ELoginCredentialType::EOS_LCT_ExchangeCode;
+	ecs.Token = nepicPassword.c_str();
+
+	EOS_Auth_LoginOptions eao = { 0 };
+	eao.ApiVersion = 1;//EOS_AUTH_LOGIN_API_LATEST;
+	//eao.ScopeFlags = 0;
+	eao.Credentials = &ecs;
+
+	static HANDLE hWait = CreateEvent(NULL, FALSE, FALSE, NULL);
+	static EOS_Auth_Token* eosToken;
+
+	static auto _EOS_Auth_CopyUserAuthToken = (decltype(&EOS_Auth_CopyUserAuthToken))GetProcAddress(eosDll, "EOS_Auth_CopyUserAuthToken");
+	auto _EOS_Auth_Login = (decltype(&EOS_Auth_Login))GetProcAddress(eosDll, "EOS_Auth_Login");
+	_EOS_Auth_Login(epicAuth, &eao, &hWait, [](const EOS_Auth_LoginCallbackInfo* info)
+	{
+		if (info->ResultCode != EOS_EResult::EOS_Success)
+		{
+			FatalError("EOS_Auth_Login failed: %d", (int)info->ResultCode);
+		}
+
+		EOS_Auth_CopyUserAuthTokenOptions cao = { 0 };
+		cao.ApiVersion = 1;
+
+		_EOS_Auth_CopyUserAuthToken(epicAuth, &cao, info->LocalUserId, &eosToken);
+
+		SetEvent(hWait);
+	});
+
+	auto _EOS_Platform_Tick = (decltype(&EOS_Platform_Tick))GetProcAddress(eosDll, "EOS_Platform_Tick");
+
+	while (WaitForSingleObject(hWait, 500) == WAIT_TIMEOUT)
+	{
+		_EOS_Platform_Tick(epicPlatform);
+ 	}
+
+	std::string s = eosToken->AccessToken;
+
+	auto j = nlohmann::json::object({{ "platform", "pcros" },
+	{ "authTicket", s } });
+
+	auto r = cpr::Post(
+		cpr::Url{ "https://rgl.rockstargames.com/api/launcher/autologinepic" },
+		cpr::Header{
+			{ { "Content-Type", "application/json; charset=utf-8" },
+			{ "Accept", "application/json" },
+			{ "X-Requested-With", "XMLHttpRequest" } } },
+		cpr::Body{
+		j.dump() },
+		cpr::VerifySsl{ false });
+
+	if (r.error)
+	{
+		bye();
+
+#ifdef IS_RDR3
+		FatalError("Error during auto-signin with ROS using Epic: %s", r.error.message);
+#endif
+
+		return;
+	}
+
+	j = nlohmann::json::parse(r.text);
+
+	if (!j["Status"].get<bool>())
+	{
+		bye();
+
+#ifdef IS_RDR3
+		FatalError("Error during Epic ROS signin: %s %s", j["Error"].value("Code", ""), j.value("Message", ""));
+#endif
+
+		return;
+	}
+
+	auto sessionKey = Botan::base64_decode(j.value("SessionKey", ""));
+
+	std::istringstream stream(j.value("Xml", ""));
+
+	boost::property_tree::ptree tree;
+	boost::property_tree::read_xml(stream, tree);
+
+	auto tick = j.value("Ticket", "");
+
+	*(uint64_t*)&blob->data[3816] = j.value("RockstarId", uint64_t(0));
+	strcpy((char*)&blob->data[2800], j.value("Ticket", "").c_str());
+	strcpy((char*)&blob->data[3312], tree.get<std::string>("Response.SessionTicket").c_str());
+	memcpy(&blob->data[0x10D8], sessionKey.data(), sessionKey.size());
+	strcpy((char*)&blob->data[8], j.value("Email", "").c_str());
+
+	j = nlohmann::json::object({
+		{ "title", "rdr2" },
+		{ "env", "prod" },
+		{ "epicPlayerJwt", s },
+		{ "playerTicket", tick },
+		{ "version", 11 },
+	});
+
+	r = cpr::Post(
+	cpr::Url{ "https://rgl.rockstargames.com/api/launcher/bindepicaccount" },
+	cpr::Header{
+	{
+	{ "Content-Type", "application/json; charset=utf-8" },
+	{ "Accept", "application/json" },
+	{ "X-Requested-With", "XMLHttpRequest" },
+	{ "Authorization", fmt::sprintf("SCAUTH val=\"%s\"", tick) },
+	} },
+	cpr::Body{
+	j.dump() },
+	cpr::VerifySsl{ false });
+
+	if (r.error)
+	{
+		bye();
+
+#ifdef IS_RDR3
+		FatalError("Error during binding of ROS to Epic: %s", r.error.message);
+#endif
+		return;
+	}
+
+	memcpy(blob->epicData, commandLineString.c_str(), commandLineString.size());
+	blob->epicSize = commandLineString.size();
+	blob->valid = true;
+
+	// resume until we can exit
+	HANDLE hEvent = OpenEvent(EVENT_MODIFY_STATE, FALSE, L"CitizenFX_GTA5_ClearedForLaunch_Epic");
+
+	if (hEvent != INVALID_HANDLE_VALUE)
+	{
+		SetEvent(hEvent);
+		CloseHandle(hEvent);
+	}
+
+	HANDLE hProcess = OpenProcess(SYNCHRONIZE, FALSE, parentPid);
+	WaitForSingleObject(hProcess, 30000);
+
+	// terminate
+	bye();
+}
+
 void ValidateSteam(int parentPid)
 {
 	static HostSharedData<ExternalROSBlob> blob("Cfx_ExtRosBlob");
@@ -272,7 +627,6 @@ void ValidateSteam(int parentPid)
 	auto tick = j.value("Ticket", "");
 
 	*(uint64_t*)&blob->data[3816] = j.value("RockstarId", uint64_t(0));
-	//*(uint64_t*)&blob->data[3816] = j.value("PlayerAccountId", uint64_t(0));
 	strcpy((char*)&blob->data[2800], j.value("Ticket", "").c_str());
 	strcpy((char*)&blob->data[3312], tree.get<std::string>("Response.SessionTicket").c_str());
 	memcpy(&blob->data[0x10D8], sessionKey.data(), sessionKey.size());
@@ -327,7 +681,7 @@ static bool InitAccountMTL()
 	if (pid == -1)
 	{
 #if defined(IS_RDR3)
-		MessageBoxW(NULL, L"Currently, you have to run the Rockstar Games Launcher to be able to run this product.", L"RedM", MB_OK | MB_ICONSTOP);
+		MessageBoxW(NULL, L"Currently, you have to run the Rockstar Games Launcher, Steam, or the Epic Games Launcher (depending on where you purchased the game) to be able to run RedM.", L"RedM", MB_OK | MB_ICONSTOP);
 #endif
 
 		return false;
@@ -438,12 +792,12 @@ void PreInitGameSpec()
 
 	accountSetUp = true;
 
-	if (wcsstr(GetCommandLineW(), L"ros:steam") != nullptr)
+	if (wcsstr(GetCommandLineW(), L"ros:steam") != nullptr || wcsstr(GetCommandLineW(), L"ros:epic") != nullptr)
 	{
 		return;
 	}
 
-	if (!InitAccountSteam() && !InitAccountMTL())
+	if (!InitAccountSteam() && !InitAccountEOS() && !InitAccountMTL())
 	{
 		TerminateProcess(GetCurrentProcess(), 0);
 	}
@@ -464,6 +818,10 @@ static HookFunction hookFunction([]()
 });
 #endif
 #else
+void ValidateEpic(int)
+{
+}
+
 void ValidateSteam(int)
 {
 }
@@ -674,9 +1032,28 @@ DWORD dwFlags)
 		{
 			lpLibFileName = va(L"%s", MakeRelativeCitPath(L"steam_api64.dll"));
 		}
+		else if (wcsstr(lpLibFileName, L"\\EOSSDK-Win64-Shipping.dll"))
+		{
+			lpLibFileName = va(L"%s", MakeRelativeGamePath(L"EOSSDK-Win64-Shipping.dll"));
+		}
 	}
 
 	return LoadLibraryExW(lpLibFileName, hFile, dwFlags);
+}
+
+static LPSTR GetCommandLineAStub()
+{
+	static char cli[65536];
+	
+	if (!cli[0])
+	{
+		static HostSharedData<ExternalROSBlob> blob("Cfx_ExtRosBlob");
+		strcpy(cli, GetCommandLineA());
+		strcat(cli, " -useEpic ");
+		strcat(cli, ToNarrow((wchar_t*)blob->epicData).c_str());
+	}
+
+	return cli;
 }
 
 static HookFunction hookFunctionSteamBlob([]()
@@ -698,6 +1075,11 @@ static HookFunction hookFunctionSteamBlob([]()
 
 			*useSteam = 1;
 			*steamAppId = strdup(va("%d", blob->steamAppId));
+		}
+		else if (blob->epicSize)
+		{
+			hook::iat("kernel32.dll", LoadLibraryExWStub, "LoadLibraryExW");
+			hook::iat("kernel32.dll", GetCommandLineAStub, "GetCommandLineA");
 		}
 	}
 #endif
